Extract Locale URLs logic into LocaleUrls concern
