+++ frontend/app/controllers/concerns/spree/locale_urls.rb
@@ -1,0 +1,23 @@
+module Spree
+  module LocaleUrls
+    extend ActiveSupport::Concern
+
+    included do
+      before_action :redirect_to_default_locale
+    end
+
+    private
+
+    def default_url_options
+      return super if locale_param.nil?
+
+      super.merge(locale: locale_param)
+    end
+
+    def redirect_to_default_locale
+      return if params[:locale].blank? || supported_locale?(params[:locale])
+
+      redirect_to url_for(request.parameters.merge(locale: nil))
+    end
+  end
+end
